Add a test to verify that translated labels work
Problem:

There was uncertainty about how to create custom labels for fields.
In https://github.com/thoughtbot/administrate/issues/451,
it was suggested that we could use I18n translations
to solve the problem.

If we accept that as the standard way to set custom labels,
we should have a test for it.

Solution:

- Add a test to verify that labels can be customized through I18n.
- Add documentation about how to customize labels

To Do:

- Update view code so the same translation definition
  applies on the show and collection pages,
  in addition to the form page.